Add CalculateController, SizeController, calculate.html, services, Ci…

+++ src/main/java/com/example/cargo/repository/CityRepository.java
@@ -3,5 +3,8 @@ package com.example.cargo.repository;
 import com.example.cargo.entity.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CityRepository extends JpaRepository<City, Integer> {
+import java.util.List;
+
+public interface CityRepository extends JpaRepository<City, Long> {
+
 }
